# docs/api/openapi.core.v1.yaml
openapi: 3.1.0
info:
  title: DELMOS Core First Slice API
  version: 0.1.0-draft
  description: >-
    Target contract for the first Core role and approval slice. Not implemented.
    Session establishment and external integrations are outside this fragment.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /projects/{project_id}/role-bindings:
    post:
      operationId: grantProjectRole
      summary: Grant a role in the project within the actor's delegation ceiling
      x-required-permission: access.grant
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantRole'
      responses:
        '201':
          description: RoleBinding and audit event recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleBinding'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Invariant' }
  /projects/{project_id}/role-bindings/{binding_id}:
    delete:
      operationId: revokeProjectRole
      summary: Revoke a project RoleBinding for subsequent requests
      x-required-permission: access.revoke
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/BindingId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Binding revoked with a durable audit event
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Invariant' }
  /projects/{project_id}/work-products:
    post:
      operationId: createWorkProduct
      summary: Create a draft Work Product with its first immutable revision
      x-required-permission: wp.create
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkProduct'
      responses:
        '201':
          description: Draft created with server-generated revision and payload hash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkProductState'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Invariant' }
  /projects/{project_id}/work-products/{work_product_id}/revisions:
    post:
      operationId: reviseWorkProduct
      summary: Write a new immutable draft revision
      x-required-permission: wp.edit
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/WorkProductId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviseWorkProduct'
      responses:
        '201':
          description: New revision created without modifying the prior revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkProductState'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Invariant' }
  /projects/{project_id}/specifications:
    post:
      operationId: createSpecification
      summary: Create a composite specification with an immutable composition manifest
      x-required-permission: wp.create
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpecification'
      responses:
        '201':
          description: Specification Work Product and first immutable revision created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecificationView'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Invariant' }
  /projects/{project_id}/specifications/{specification_id}:
    get:
      operationId: getSpecification
      summary: Read an immutable composite specification manifest
      x-required-permission: wp.read
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: specification_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Specification and exact composition manifest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecificationView'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
  /projects/{project_id}/specifications/{specification_id}/revisions:
    post:
      operationId: reviseSpecification
      summary: Write a new immutable specification revision and manifest
      x-required-permission: wp.edit
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: specification_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviseSpecification'
      responses:
        '201':
          description: New immutable specification revision created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecificationView'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Invariant' }
...
  /projects/{project_id}/plan/apply:
    post:
      operationId: applyApprovedPlan
      summary: Apply a separately approved PLAN-001 revision
      x-required-permission: plan.apply
      description: >-
        The target revision must already carry a positive approval recorded
        through the review workflow (ADR-009 §5); otherwise the call fails with
        422 plan_not_approved. Approval is checked per revision, not by the work
        product status: once approved the plan stays approved, so a newer
        unapproved revision would otherwise slip into effect unnoticed.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevisionAction'
      responses:
        '200':
          description: Effective plan revision and configuration generation updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanApplied'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Invariant' }
...
  /projects/{project_id}/trace-links:
    post:
      operationId: createTraceLink
      summary: Create a typed traceability link between project work products
      x-required-permission: trace.create
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTraceLink'
      responses:
        '201':
          description: Traceability link created and audited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceLink'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Invariant' }
    get:
      operationId: listTraceLinks
      summary: List project traceability links (Change Impact review list, GRP-03)
      x-required-permission: wp.read
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: suspect
          in: query
          required: false
          description: When "true", return only links flagged is_suspect after a source revision.
          schema: { type: string, enum: ['true', 'false'] }
      responses:
        '200':
          description: Traceability links, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TraceLink' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
  /projects/{project_id}/trace-links/{trace_link_id}/acknowledge:
    post:
      operationId: acknowledgeTraceLink
      summary: Clear the is_suspect flag after reviewer confirmation (GRP-03)
      x-required-permission: trace.create
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: trace_link_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { description: Suspect flag cleared or already clear (idempotent) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
  /projects/{project_id}/work-products/{work_product_id}/similar:
    get:
      operationId: findSimilarWorkProducts
      summary: Hybrid semantic search for near-duplicate requirements/tests (VEC-01..03)
      x-required-permission: wp.read
      description: >-
        Compares the work product's current revision embedding against other
        embedded revisions in the same project via pgvector cosine distance.
        Only vectors of the current embedding model are matched; permission
        scope is applied as a project_id SQL filter, not post-filtering.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: work_product_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Similar work products ordered by descending cosine similarity
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SimilarWorkProduct' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
  /projects/{project_id}/traceability/{source_id}:
    get:
      operationId: traverseTraceability
      summary: Traverse the project traceability graph from a work product
      x-required-permission: wp.read
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: source_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Bounded recursive traceability path
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TracePathEntry' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
  /system/automation/status:
    get:
      operationId: getAutomationStatus
      summary: Observe the automation engine outbox queue (ADR-007, SPEC-04)
      x-required-permission: system.observe
      description: >-
        Delivery counts by status and the dead-letter list for administrator
        monitoring (RUNBOOK.md, ADM-007). Not a project-scoped endpoint.
      responses:
        '200':
          description: Delivery status counts and dead-letter entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationStatus'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { description: Automation engine not running }
...
  /projects/{project_id}/economics/earned-value:
    get:
      operationId: getEarnedValue
      summary: Earned value metrics of the project (ДСТУ ISO 21508, SWR-41)
      x-required-permission: economics.read
      description: >-
        Earned value management metrics per ISO 21508 (ADR-011). Planned value
        is time-phased linearly across the phase calendar; earned value uses the
        0/100 rule over approved deliverables linked to the phase (ISO 21511).
        Monetary values are decimal strings computed in PostgreSQL NUMERIC and
        are never converted to floating point. Ratio fields are null when the
        denominator is zero: absence of data must not be read as zero.
      parameters:
        - name: project_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: as_of
          in: query
          required: false
          description: Control date (YYYY-MM-DD). Defaults to the current date.
          schema: { type: string, format: date }
      responses:
        '200':
          description: Earned value snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EarnedValueSnapshot'
        '400': { description: Malformed as_of parameter }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: No approved cost baseline for the project }
        '422': { description: Incomplete cost data (missing labor rate or mixed currencies) }
  /projects/{project_id}/phases/{phase_key}/transition:
    post:
      operationId: transitionPhase
      summary: Move a project phase through its lifecycle (SHR-13, ADR-006)
      x-required-permission: plan.apply
      description: >-
        Transitions a phase not_started -> active -> completed. Backward and
        skipping transitions are rejected: a completed phase is already
        reflected in earned value figures and reopening it would rewrite the
        totals the gate relied on. Economic facts (actual cost, funding limit,
        CPI) are computed inside the same transaction as the status change, so
        no expense can be recorded between the check and the write.
        rule.economics.funding_limit_gate vetoes closure when actual cost
        exceeds the funding limit; rule.economics.cost_variance_alert only
        records an audit warning.
      parameters:
        - name: project_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: phase_key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_status]
              properties:
                target_status:
                  type: string
                  enum: [active, completed]
      responses:
        '200':
          description: Phase transitioned
          content:
            application/json:
              schema:
                type: object
                required: [phase_key, from_status, to_status]
                properties:
                  phase_key: { type: string }
                  from_status: { type: string }
                  to_status: { type: string }
        '400': { description: Malformed body }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Phase not found in project }
        '422': { description: Invalid transition or blocked by a mandatory rule }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: delmos_session
      description: Server-side HttpOnly session; its creation is outside this fragment.
  parameters:
...
    ProjectId:
      in: path
      name: project_id
      required: true
      schema: { type: string, format: uuid }
...
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string, minLength: 1 }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: No valid session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Missing permission for a visible action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Hidden:
      description: Resource is missing or not visible in this scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Stale row version or idempotency key reused with different payload
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Invariant:
      description: Validation, SoD, quorum or lifecycle rule failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    GrantRole:
      type: object
      additionalProperties: false
      required: [user_id, role_key, reason]
      properties:
        user_id: { type: string, format: uuid }
        role_key:
          type: string
          enum: [viewer, engineer, reviewer, approver, auditor, project_manager]
        reason: { type: string, minLength: 1 }
        expires_at: { type: string, format: date-time }
    RoleBinding:
      type: object
      required: [id, user_id, role_key, project_id]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        role_key: { type: string }
        project_id: { type: string, format: uuid }
...